[Feat] 사용자 생성 및 조회 (김세현)

+++ src/main/java/com/chaebbi/community/domain/Images.java
@@ -14,7 +14,7 @@ import javax.persistence.*;
 public class Images {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idx;
+    private Long idx;
 
     private int postIdx;
     private String imgUrl;
